% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaid.R
\name{chaid}
\alias{chaid}
\title{Constructs a decision tree using the CHAID algorithm.}
\usage{
chaid(
  formula,
  data,
  subset,
  weights,
  na.action = na.omit,
  control = chaid_control()
)
}
\arguments{
\item{formula}{A formula expression. The left-hand-side (response) should be
a factor. The right-hand-side should be a series of factor variables
separated by \code{+}.}

\item{data}{a data frame containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be \code{NULL} or a numeric vector.}

\item{na.action}{a function which indicates what should happen when the data
contains \code{NA}s. The default is \code{na.omit}.}

\item{control}{hyper parameters of the algorithm as returned by
\code{\link{chaid_control}}.}
}
\value{
An object of class \code{constparty}, see package
    \code{\link[partykit]{party}}.
}
\description{
The CHAID-algorithm has 5 steps. Somewhat counterintuitively, we start with
step 5.
}
\details{
Step 1: calculates the cross-tabs for each predictor variable with the
outcome variable in turn and performs step 2 and step 3.
Step 2: Find the two predictors which show the least significant differing
crosstab with the outcome. If the p-value exceeds the critical value, the
categories are merged.
Step 3: For groups of 3 or more categories, split groups apart if the
significance of such a split would be above some threshold.
Step 4: computes the significance of each optimally grouped predictor. If the
significance of splitting using this predictor is above some threshold,
perform the split.
Step 5: Recursively construct the tree by applying steps 1 and 4 to each
unevaluated partition of the data, until no more partitions are formed (i.e.
the tree is complete).
}
\examples{
data(adult.2000, package = "flipExampleData")
chaid(income ~ education + occupation + race, data = adult.2000)
}
\references{
G. V. Kass (1980). An Exploratory Technique for Investigating
Large Quantities of Categorical Data. \emph{Applied Statistics},
\bold{29}(2), 119--127.
}
