% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaid.R
\name{chaid_control}
\alias{chaid_control}
\title{Default hyperparameters for the CHAID algorithm.}
\usage{
chaid_control(
  alpha2 = 0.05,
  alpha3 = -1,
  alpha4 = 0.05,
  minsplit = 20,
  minbucket = 7,
  minprob = 0.01,
  stump = FALSE,
  maxheight = -1
)
}
\arguments{
\item{alpha2}{Level of significance used for merging of predictor categories
(step 2).}

\item{alpha3}{- If set to a positive value < 1, level of significance used
for the the splitting of former merged categories of the predictor (step
3). Otherwise, step 3 is omitted (the default).}

\item{alpha4}{- Level of significance used for splitting of a node in the
most significant predictor (step 5).}

\item{minsplit}{- Number of observations in splitted response at which no
further split is desired.}

\item{minbucket}{- Minimum number of observations in terminal nodes.}

\item{minprob}{- Mininimum frequency of observations in terminal nodes. (What
does this mean???)}

\item{stump}{- if TRUE, limits height to 1 (decision stump/1 rule learning)}

\item{maxheight}{- max height of tree. no limit if -1}
}
\description{
Overriden if the user specifies the 'control' parameter \code{\link{chaid}}.
}
