% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttree.R
\name{CheckDataForTextTree}
\alias{CheckDataForTextTree}
\title{\code{CheckDataForTextTree} Check that there are enough cases to compute the
tree from the term document matrix.}
\usage{
CheckDataForTextTree(
  data,
  weights = NULL,
  subset = NULL,
  missing = "Exclude cases with missing data"
)
}
\arguments{
\item{data}{A data frame where the first column is the outcome variable and the subsequent
columns are terms from the term document matrix.}

\item{weights}{An option vector containing sampling weights for each row of \code{data}.}

\item{subset}{An option vector describing the subset of rows from \code{data} that are to be used to generate the tree.}

\item{missing}{How missing data is to be treated in the
regression. Options are: \code{"Error if missing data"},
\code{"Exclude cases with missing data"},
\code{"Use partial data"},and
\code{"Imputation (replace missing values with estimates)"}.}
}
\description{
\code{CheckDataForTextTree} Check that there are enough cases to compute the
tree from the term document matrix.
}
