% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttree.R
\name{CreateTextTree}
\alias{CreateTextTree}
\title{\code{CreateTextTree} Generate a predictive tree from an outcome variable and a tree.}
\usage{
CreateTextTree(tree, outcome.variable, original.text, transformed.text,
  output = "Sankey")
}
\arguments{
\item{tree}{A tree object.}

\item{outcome.variable}{The dependent variable used when creating \code{tree}.}

\item{original.text}{A vector containing the source text that was used to create the tree.}

\item{transformed.text}{A vector containing the processed text.}

\item{output}{A string indicating how the result should be printed. Options
are \code{"Sankey"}, \code{"Tree"}, \code{"Text"}, or \code{"Table"}. The first three
match the output options of \code{\link{CART}}, and "Table" prints an HTML widget table
which shows the outcome variable next to the text.
#' @inheritParams CART}
}
\description{
\code{CreateTextTree} Generate a predictive tree from an outcome variable and a tree.
}

