% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{predict.CART}
\alias{predict.CART}
\title{predict.CART}
\usage{
\method{predict}{CART}(object, seed = 1232, ...)
}
\arguments{
\item{object}{The \code{CART} object whose values are to be predicted.}

\item{seed}{A random number seed to ensure stability of predictions.}
}
\description{
predict.CART
}

