% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{Probabilities.CART}
\alias{Probabilities.CART}
\title{Probabilities.CART}
\usage{
Probabilities.CART(object)
}
\arguments{
\item{object}{The \code{CART} object whose values are to be predicted.}
}
\description{
Probabilities.CART
}

