% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{treeFrameToList}
\alias{treeFrameToList}
\title{\code{treeFrameToList} Converts a \code{tree} into a format usable in
a sankeytree.}
\usage{
treeFrameToList(
  frame,
  xlevels,
  model,
  assigned,
  labels,
  max.tooltip.length = 150,
  numeric.distribution = TRUE,
  custom.color = "default",
  num.color.div = 101,
  const.bin.size = TRUE
)
}
\arguments{
\item{frame}{The tree frame.}

\item{xlevels}{Levels of depedent variables that are factors.}

\item{model}{Data used by the model.}

\item{assigned}{The nodes that the cases have been assigned.}

\item{labels}{A vector of variable labels, named by the variable names.}

\item{max.tooltip.length}{The maximum length of the tooltip (determines the
scale of the tree).}

\item{numeric.distribution}{Outputs additional diagnostics in the tooltip.}

\item{custom.color}{Determines the colors of tree branches; if \code{"default"}
generates tree colors using Q colors; if \code{"sankey"} use colors provided by sankeyTree package;
or can provide a vector of hex strings as RGB values, for example \code{custom.color = c("#aabbcc","#123456")}.}

\item{num.color.div}{positive integer in the range [2,inf]. Controls the
color resolution of the tree. A higher value gives a smoother color
transition.}

\item{const.bin.size}{logical; if \code{true}, each color spans an equal step
of y-value or an equal number of points.}
}
\description{
\code{treeFrameToList} Converts a \code{tree} into a format usable in
a sankeytree.
}
